Player Movement Attacks
Some attacks will push the player in a given direction, The player state will need to change its root at some points during run time.
Example: Grounded -> Airborne instantly when the attack adds force to the player,
Signal from the animation of the attack signals the player to add force and to switch the root state if applicable.